bing-search-agent/requirements.txt
fastmcp
azure-identity>=1.0
azure-ai-projects>=1.0.0b11
python-dotenv>=1.0.0
aiohttp>=3.8.0